Bugfix: Unexpected disconnect after enabling CCCD
This doesn't look as a bug in library. So this is a small hack that
works for some Android devices, but doesn't harm dfu process on other
devices at all.

E.g. without this hack, dfu process can't finish successfully on HTC M9,
running Android 5.1.1. It disconnets two operations (calling two writeOpCode
methods) after enabling CCCD, with one of this statuses: 133, 14 or 8.