Mark position after execute for CRC retry
I've noticed when using in a noisy environment if I ever see a CRC error
all of the subsequent retries fail and the download fails.

I noticed if I print the CRC values expected they change on each retry
attempt.  This seems wrong, since I believe the intent is just to re-send
the last chunk.

I suspect there's a mark() call missing in the success transfer case,
and the position of the last successful create->execute transfer should
be saved. The next retry transfer should only reset to this mark position
and not the beginning of the file.

If I add a call to mFirwmareStream.mark in the CRC success case after
writeExecute() completes I am able to successfully recover from CRC errors.